% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pOrg.R
\name{function.R_txt}
\alias{function.R_txt}
\title{function.R}
\usage{
function.R_txt()
}
\arguments{
\item{Values}{}
}
\value{
Text that will be used to for the template

"
}
\description{
This function contains the text that is used for the R file function.R.
This file is used to hold custom functions used in the project.
}
