% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pOrg.R
\name{pOrg}
\alias{pOrg}
\title{Project organizer}
\usage{
pOrg(
  path = getwd(),
  folder_lv1 = c("project_documents", "data", "code", "output", "report"),
  folder_lv2 = c("data/original", "data/derived", "output/plots", "output/figures",
    "output/tables", "code/exploratory", "code/final_code", "code/working_code"),
  include_readme = T,
  include_function = T,
  include_rmd_template = T,
  folder_files = "code"
)
}
\arguments{
\item{path}{character containing the path to use for the project folder structure.
By default uses the working directory returned by getwd().}

\item{folder_lv1}{character vector with the names of the upper level folders}

\item{folder_lv2}{character vector with the names of the lower level folders}

\item{include_readme}{boolean whether to include README file}

\item{include_function}{boolean whether to include the function.R file}

\item{include_rmd_template}{boolean whether to include the Rmd template file}

\item{folder_files}{character with the name of the folder from folder_lv1 or folder_lv2
for placing the function.R and Rmd template files.}
}
\description{
This function creates the folder structure given the specified path.
}
\details{
The project structure by default is as follow:

\enumerate{
  \item{\strong{README.txt}}
  \item{\strong{code}}
  \enumerate{
   \item{\strong{exploratory: }}{code that is exploratory in nature}
   \item{\strong{final_code: }}{final code that reproduced the final analysis}
   \item{\strong{working_code: }}{temporary code during work in progress}
   \item{\strong{functions.R: }}{the custom functions used in the project}
   }
  \item{\strong{data}}
  \enumerate{
   \item{\strong{derived: }}{data generated as a product of the analysis}
   \item{\strong{original: }}{original data}
   }
  \item{\strong{output}}
  \enumerate{
   \item{\strong{figures: }}{composited figures}
   \item{\strong{plots: }}{individual plots}
   \item{\strong{tabels: }}{individual tabels}
   }
  \item{\strong{project_documents}}
  \item{\strong{report}}
}
}
