% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pOrg.R
\name{readme_txt}
\alias{readme_txt}
\title{README file text}
\usage{
readme_txt()
}
\value{
Text that will be used to for the template
}
\description{
This function contains the text that is used for creating the README.txt file
}
