% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pOrg.R
\name{Rmd_template}
\alias{Rmd_template}
\title{RMarkdown template text}
\usage{
Rmd_template()
}
\value{
Text that will be used to for the template
}
\description{
This function contains the text that is used as a template for the Rmd files.
}
