% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pOrg.R
\name{set_pOrg_txt}
\alias{set_pOrg_txt}
\title{set project files template}
\usage{
set_pOrg_txt(txt, file_template, set_defualt = F)
}
\arguments{
\item{txt}{character with the text that will be used as a template for the specified file_template}

\item{file_template}{name of the file for which to set the template. The options are "README_txt", "
function.R_txt","Rmd_template_txt"}

\item{set_default}{whether to set the files to the original installed with the package.
If set to true, txt will be ignored}
}
\description{
This function sets the text that is used for creating the README.txt,
function.R and Rmd template files
}
